use navigate with high order component
karena pakai router react dom v6, history sudah deprecated
