+++ my-course/src/pages/MyCourseClass.js
@@ -1,0 +1,51 @@
+import React, { Component } from 'react';
+// import PropTypes from 'prop-types';
+
+
+class MyCourseClass extends Component {
+    constructor(props) {
+        super(props);
+        this.state = []; 
+    }
+
+    // componentWillMount() {
+    // }
+
+    componentDidMount() {
+        
+    }
+
+    componentWillReceiveProps(nextProps) {
+
+    }
+
+    shouldComponentUpdate(nextProps, nextState) {
+
+    }
+
+    componentWillUpdate(nextProps, nextState) {
+
+    }
+
+    componentDidUpdate(prevProps, prevState) {
+
+    }
+
+    componentWillUnmount() {
+
+    }
+
+    render() {
+        return (
+            <div>
+
+            </div>
+        );
+    }
+}
+
+// MyCourseClass.propTypes = {
+
+// };
+
+export default MyCourseClass;
